Combine duplicate test case into one file
