+++ tests/logic_adapter_tests/best_match_integration_tests/test_levenshtein_distance.py
@@ -21,14 +21,6 @@ class BestMatchLevenshteinDistanceTestCase(TestCase):
         # Add a mock chatbot to the logic adapter
         self.adapter.set_chatbot(MockChatBot())
 
-    def test_no_choices(self):
-        self.adapter.chatbot.storage.filter = MagicMock(return_value=[])
-
-        statement = Statement("What is your quest?")
-
-        with self.assertRaises(BestMatch.EmptyDatasetException):
-            self.adapter.get(statement)
-
     def test_get_closest_statement(self):
         """
         Note, the content of the in_response_to field for each of the
